Feat: Melhoria no carrinho, diminuindo fontes

+++ db/schema.rb
@@ -67,7 +67,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
   end
 
   create_table "aulas", force: :cascade do |t|
-    t.integer "modulo_id", null: false
+    t.bigint "modulo_id", null: false
     t.string "titulo", null: false
     t.string "description"
     t.integer "duration"
@@ -106,8 +106,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
   end
 
   create_table "curso_words", force: :cascade do |t|
-    t.integer "curso_id", null: false
-    t.integer "word_id", null: false
+    t.bigint "curso_id", null: false
+    t.bigint "word_id", null: false
     t.index ["curso_id"], name: "index_curso_words_on_curso_id"
     t.index ["word_id"], name: "index_curso_words_on_word_id"
   end
@@ -134,8 +134,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
   end
 
   create_table "cursos_words", id: false, force: :cascade do |t|
-    t.integer "curso_id", null: false
-    t.integer "word_id", null: false
+    t.bigint "curso_id", null: false
+    t.bigint "word_id", null: false
     t.index ["curso_id"], name: "index_cursos_words_on_curso_id"
     t.index ["word_id"], name: "index_cursos_words_on_word_id"
   end
@@ -168,8 +168,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
     t.decimal "preço_final"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
-    t.integer "address_origem_id", null: false
-    t.integer "address_destino_id", null: false
+    t.bigint "address_origem_id", null: false
+    t.bigint "address_destino_id", null: false
     t.index ["address_destino_id"], name: "index_freights_on_address_destino_id"
     t.index ["address_origem_id"], name: "index_freights_on_address_origem_id"
   end
@@ -186,7 +186,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
   end
 
   create_table "modulos", force: :cascade do |t|
-    t.integer "curso_id", null: false
+    t.bigint "curso_id", null: false
     t.string "titulo", null: false
     t.string "description"
     t.integer "position"
@@ -232,7 +232,7 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
   end
 
   create_table "texts", force: :cascade do |t|
-    t.integer "word_id", null: false
+    t.bigint "word_id", null: false
     t.text "content"
     t.index ["word_id"], name: "index_texts_on_word_id"
   end
@@ -250,8 +250,8 @@ ActiveRecord::Schema[8.0].define(version: 2025_07_27_155433) do
   end
 
   create_table "user_cursos", force: :cascade do |t|
-    t.integer "user_id", null: false
-    t.integer "curso_id", null: false
+    t.bigint "user_id", null: false
+    t.bigint "curso_id", null: false
     t.integer "progresso", default: 0, null: false
     t.boolean "concluido", default: false, null: false
     t.datetime "data_inicio"
